+++ pkg/types/update_params.go
@@ -19,4 +19,5 @@ type UpdateParams struct {
 	LifecycleUID     int           // Default UID for lifecycle hooks.
 	LifecycleGID     int           // Default GID for lifecycle hooks.
 	NoSelfUpdate     bool          // Skip self-update of Watchtower if true.
+	CPUCopyMode      string        // CPU copy mode for container recreation.
 }